assets: images moved
